# Add support for client indexes