--- a/cache/cache.go
+++ b/cache/cache.go
@@ -25,6 +25,7 @@ const (
 	deleteEvent     = "delete"
 	bufferSize      = 65536
 	columnDelimiter = ","
+	keyDelimiter    = "|"
 )
 
 // ErrCacheInconsistent is an error that can occur when an operation
@@ -68,32 +69,77 @@ func NewIndexExistsError(table string, value interface{}, index string, new, exi
 }
 
 // map of unique values to uuids
-type valueToUUID map[interface{}]string
+type valueToUUIDs map[interface{}]uuidset
 
-// map of column name(s) to a unique values, to UUIDs
-type columnToValue map[index]valueToUUID
+// map of column name(s) to unique values, to UUIDs
+type columnToValue map[index]valueToUUIDs
 
 // index is the type used to implement multiple cache indexes
 type index string
 
-// columns returns the columns that conform the index
-func (i index) columns() []string {
-	return strings.Split(string(i), columnDelimiter)
+// indexType is the type of index
+type indexType uint
+
+const (
+	schemaIndexType indexType = iota
+	clientIndexType
+)
+
+// indexSpec contains details about an index
+type indexSpec struct {
+	index     index
+	columns   []model.ColumnKey
+	indexType indexType
+}
+
+func (s indexSpec) isClientIndex() bool {
+	return s.indexType == clientIndexType
+}
+
+func (s indexSpec) isSchemaIndex() bool {
+	return s.indexType == schemaIndexType
 }
 
 // newIndex builds a index from a list of columns
-func newIndex(columns ...string) index {
+func newIndexFromColumns(columns ...string) index {
 	return index(strings.Join(columns, columnDelimiter))
+}
+
+// newIndexFromColumnKeys builds a index from a list of column keys
+func newIndexFromColumnKeys(columns ...model.ColumnKey) index {
+	// RFC 7047 says that Indexes is a [<column-set>] and "Each <column-set> is a set of
+	// columns whose values, taken together within any given row, must be
+	// unique within the table". We'll store the column names, separated by comma
+	// as we'll assume (RFC is not clear), that comma isn't valid in a <id>
+	columnIndexes := make([]string, len(columns))
+	for i, column := range columns {
+		if column.Key != nil {
+			columnIndexes[i] = fmt.Sprintf("%s%s%v", column.Column, keyDelimiter, column.Key)
+		} else {
+			columnIndexes[i] = column.Column
+		}
+	}
+	return newIndexFromColumns(columnIndexes...)
+}
+
+// newColumnKeysFromColumns builds a list of column keys from a list of columns
+func newColumnKeysFromColumns(columns ...string) []model.ColumnKey {
+	columnKeys := make([]model.ColumnKey, len(columns))
+	for i, column := range columns {
+		columnKeys[i] = model.ColumnKey{Column: column}
+	}
+	return columnKeys
 }
 
 // RowCache is a collections of Models hashed by UUID
 type RowCache struct {
-	name     string
-	dbModel  model.DatabaseModel
-	dataType reflect.Type
-	cache    map[string]model.Model
-	indexes  columnToValue
-	mutex    sync.RWMutex
+	name       string
+	dbModel    model.DatabaseModel
+	dataType   reflect.Type
+	cache      map[string]model.Model
+	indexSpecs []indexSpec
+	indexes    columnToValue
+	mutex      sync.RWMutex
 }
 
 // rowByUUID returns one model from the cache by UUID. Caller must hold the row
@@ -112,33 +158,70 @@ func (r *RowCache) Row(uuid string) model.Model {
 	return r.rowByUUID(uuid)
 }
 
-// RowByModel searches the cache using indexes for a provided model. Returns the
-// found Model along with its UUID.
-func (r *RowCache) RowByModel(m model.Model) (string, model.Model) {
+func (r *RowCache) rowsByModel(m model.Model, useClientIndexes bool) map[string]model.Model {
 	r.mutex.RLock()
 	defer r.mutex.RUnlock()
 	if reflect.TypeOf(m) != r.dataType {
-		return "", nil
+		return nil
 	}
 	info, _ := r.dbModel.NewModelInfo(m)
 	field, err := info.FieldByColumn("_uuid")
 	if err != nil {
-		return "", nil
+		return nil
 	}
 	uuid := field.(string)
 	if uuid != "" {
-		return uuid, r.rowByUUID(uuid)
+		row := r.rowByUUID(uuid)
+		if row != nil {
+			return map[string]model.Model{uuid: row}
+		}
 	}
-	for index, vals := range r.indexes {
-		val, err := valueFromIndex(info, index)
+	// indexSpecs are ordered, schema indexes go first, then client indexes
+	for _, indexSpec := range r.indexSpecs {
+		if indexSpec.isClientIndex() && !useClientIndexes {
+			// Given the ordered indexSpecs, we can break here if we reach the
+			// first client index
+			break
+		}
+		val, err := valueFromIndex(info, indexSpec.columns)
 		if err != nil {
 			continue
 		}
-		if uuid, ok := vals[val]; ok {
-			return uuid, r.rowByUUID(uuid)
+		vals := r.indexes[indexSpec.index]
+		uuids, ok := vals[val]
+		if !ok || uuids.empty() {
+			continue
 		}
+		results := make(map[string]model.Model, len(uuids))
+		for uuid := range uuids {
+			results[uuid] = r.rowByUUID(uuid)
+		}
+		return results
+	}
+	return nil
+}
+
+// RowByModel searches the cache by UUID and schema indexes. UUID search is
+// performed first. Schema indexes are evaluated in turn by the same order with
+// which they are defined in the schema. First found Model is returned along
+// with its UUID. An empty string and nil is returned if no Model is found.
+func (r *RowCache) RowByModel(m model.Model) (string, model.Model) {
+	models := r.rowsByModel(m, false)
+	for uuid, model := range models {
+		return uuid, model
 	}
 	return "", nil
+}
+
+// RowsByModel searches the cache by UUID, schema indexes and client indexes.
+// UUID search is performed first. Schema indexes are evaluated next in turn by
+// the same order with which they are defined in the schema. Finnally, client
+// indexes are evaluated in turn by the same order with which they are defined
+// in the client DB model. First found Models are returned, which might be more
+// than 1 if they were found through a client index since in that case
+// uniqueness is not enforced. Nil is returned if no Model is found.
+func (r *RowCache) RowsByModel(m model.Model) map[string]model.Model {
+	return r.rowsByModel(m, true)
 }
 
 // Create writes the provided content to the cache
@@ -155,25 +238,31 @@ func (r *RowCache) Create(uuid string, m model.Model, checkIndexes bool) error {
 	if err != nil {
 		return err
 	}
-	newIndexes := newColumnToValue(r.dbModel.Schema.Table(r.name).Indexes)
-	for index, vals := range r.indexes {
-		val, err := valueFromIndex(info, index)
+	newIndexes := r.newIndexes()
+	for _, indexSpec := range r.indexSpecs {
+		index := indexSpec.index
+		val, err := valueFromIndex(info, indexSpec.columns)
 		if err != nil {
 			return err
 		}
 
-		if existing, ok := vals[val]; ok && checkIndexes {
-			return NewIndexExistsError(r.name, val, string(index), uuid, existing)
+		vals := r.indexes[index]
+		if existing, ok := vals[val]; ok && !existing.empty() && checkIndexes && indexSpec.isSchemaIndex() {
+			return NewIndexExistsError(r.name, val, string(index), uuid, existing.getAny())
 		}
 
-		newIndexes[index][val] = uuid
+		uuidset := newUUIDSet(uuid)
+		if indexSpec.isSchemaIndex() {
+			newIndexes[index][val] = uuidset
+		} else {
+			newIndexes[index][val] = addUUIDSet(r.indexes[index][val], uuidset)
+		}
 	}
 
 	// write indexes
 	for k1, v1 := range newIndexes {
-		vals := r.indexes[k1]
 		for k2, v2 := range v1 {
-			vals[k2] = v2
+			r.indexes[k1][k2] = v2
 		}
 	}
 	r.cache[uuid] = model.Clone(m)
@@ -196,17 +285,16 @@ func (r *RowCache) Update(uuid string, m model.Model, checkIndexes bool) error {
 	if err != nil {
 		return err
 	}
-	indexes := r.dbModel.Schema.Table(r.name).Indexes
-	newIndexes := newColumnToValue(indexes)
-	oldIndexes := newColumnToValue(indexes)
+	newIndexes := r.newIndexes()
 	var errs []error
-	for index, vals := range r.indexes {
+	for _, indexSpec := range r.indexSpecs {
+		index := indexSpec.index
 		var err error
-		oldVal, err := valueFromIndex(oldInfo, index)
+		oldVal, err := valueFromIndex(oldInfo, indexSpec.columns)
 		if err != nil {
 			return err
 		}
-		newVal, err := valueFromIndex(newInfo, index)
+		newVal, err := valueFromIndex(newInfo, indexSpec.columns)
 		if err != nil {
 			return err
 		}
@@ -218,34 +306,37 @@ func (r *RowCache) Update(uuid string, m model.Model, checkIndexes bool) error {
 		// old and new values are NOT the same
 
 		// check that there are no conflicts
-		if conflict, ok := vals[newVal]; ok && checkIndexes && conflict != uuid {
+		vals := r.indexes[index]
+		if existing, ok := vals[newVal]; ok && indexSpec.isSchemaIndex() && checkIndexes && !existing.empty() && !existing.has(uuid) {
 			errs = append(errs, NewIndexExistsError(
 				r.name,
 				newVal,
 				string(index),
 				uuid,
-				conflict,
+				existing.getAny(),
 			))
 		}
 
-		newIndexes[index][newVal] = uuid
-		oldIndexes[index][oldVal] = ""
+		uuidset := newUUIDSet(uuid)
+		if indexSpec.isSchemaIndex() {
+			newIndexes[index][newVal] = uuidset
+			newIndexes[index][oldVal] = nil
+		} else {
+			newIndexes[index][newVal] = addUUIDSet(r.indexes[index][newVal], uuidset)
+			newIndexes[index][oldVal] = substractUUIDSet(r.indexes[index][oldVal], uuidset)
+		}
 	}
 	if len(errs) > 0 {
 		return fmt.Errorf("%+v", errs)
 	}
 	// write indexes
 	for k1, v1 := range newIndexes {
-		vals := r.indexes[k1]
 		for k2, v2 := range v1 {
-			vals[k2] = v2
-		}
-	}
-	// delete old indexes
-	for k1, v1 := range oldIndexes {
-		vals := r.indexes[k1]
-		for k2 := range v1 {
-			delete(vals, k2)
+			if len(v2) == 0 {
+				delete(r.indexes[k1], k2)
+			} else {
+				r.indexes[k1][k2] = v2
+			}
 		}
 	}
 	r.cache[uuid] = model.Clone(m)
@@ -257,22 +348,30 @@ func (r *RowCache) IndexExists(row model.Model) error {
 	if err != nil {
 		return err
 	}
-	uuid, err := info.FieldByColumn("_uuid")
+	field, err := info.FieldByColumn("_uuid")
 	if err != nil {
 		return nil
 	}
-	for index, vals := range r.indexes {
-		val, err := valueFromIndex(info, index)
+	uuid := field.(string)
+	for _, indexSpec := range r.indexSpecs {
+		if !indexSpec.isSchemaIndex() {
+			// Given the ordered indexSpecs, we can break here if we reach the
+			// first non schema index
+			break
+		}
+		index := indexSpec.index
+		val, err := valueFromIndex(info, indexSpec.columns)
 		if err != nil {
 			continue
 		}
-		if existing, ok := vals[val]; ok && existing != uuid.(string) {
+		vals := r.indexes[index]
+		if existing := vals[val]; !existing.empty() && !existing.has(uuid) {
 			return NewIndexExistsError(
 				r.name,
 				val,
 				string(index),
-				uuid.(string),
-				existing,
+				uuid,
+				existing.getAny(),
 			)
 		}
 	}
@@ -291,16 +390,22 @@ func (r *RowCache) Delete(uuid string) error {
 	if err != nil {
 		return err
 	}
-	for index, vals := range r.indexes {
-		oldVal, err := valueFromIndex(oldInfo, index)
+	for _, indexSpec := range r.indexSpecs {
+		index := indexSpec.index
+		oldVal, err := valueFromIndex(oldInfo, indexSpec.columns)
 		if err != nil {
 			return err
 		}
 		// only remove the index if it is pointing to this uuid
 		// otherwise we can cause a consistency issue if we've processed
 		// updates out of order
-		if vals[oldVal] == uuid {
-			delete(vals, oldVal)
+		vals := r.indexes[index]
+		existing, ok := vals[oldVal]
+		if ok {
+			existing.remove(uuid)
+			if len(existing) == 0 {
+				delete(vals, oldVal)
+			}
 		}
 	}
 	delete(r.cache, uuid)
@@ -359,8 +464,8 @@ func (r *RowCache) RowsByCondition(conditions []ovsdb.Condition) (map[string]mod
 					results[rowUUID] = row
 				}
 			}
-		} else if index, err := r.Index(condition.Column); err == nil {
-			for k, rowUUID := range index {
+		} else if index, err := r.indexForColumns(condition.Column); err == nil {
+			for k, uuids := range index {
 				tSchema := schema.Columns[condition.Column]
 				nativeValue, err := ovsdb.OvsToNative(tSchema, condition.Value)
 				if err != nil {
@@ -371,8 +476,10 @@ func (r *RowCache) RowsByCondition(conditions []ovsdb.Condition) (map[string]mod
 					return nil, err
 				}
 				if ok {
-					row := r.Row(rowUUID)
-					results[rowUUID] = row
+					for uuid := range uuids {
+						row := r.Row(uuid)
+						results[uuid] = row
+					}
 				}
 			}
 		} else {
@@ -410,12 +517,27 @@ func (r *RowCache) Len() int {
 	return len(r.cache)
 }
 
-func (r *RowCache) Index(columns ...string) (map[interface{}]string, error) {
+func (r *RowCache) Index(columns ...string) (map[interface{}][]string, error) {
 	r.mutex.RLock()
 	defer r.mutex.RUnlock()
-	index, ok := r.indexes[newIndex(columns...)]
+	spec := newIndexFromColumns(columns...)
+	index, ok := r.indexes[spec]
 	if !ok {
-		return nil, fmt.Errorf("%s is not an index", index)
+		return nil, fmt.Errorf("%v is not an index", columns)
+	}
+	dbIndex := make(map[interface{}][]string, len(index))
+	for k, v := range index {
+		dbIndex[k] = v.list()
+	}
+	return dbIndex, nil
+}
+
+func (r *RowCache) indexForColumns(columns ...string) (valueToUUIDs, error) {
+	r.mutex.RLock()
+	defer r.mutex.RUnlock()
+	index, ok := r.indexes[newIndexFromColumns(columns...)]
+	if !ok {
+		return nil, fmt.Errorf("%v is not an index", columns)
 	}
 	return index, nil
 }
@@ -742,29 +864,49 @@ func (t *TableCache) Run(stopCh <-chan struct{}) {
 // newRowCache creates a new row cache with the provided data
 // if the data is nil, and empty RowCache will be created
 func newRowCache(name string, dbModel model.DatabaseModel, dataType reflect.Type) *RowCache {
+	schemaIndexes := dbModel.Schema.Table(name).Indexes
+	clientIndexes := dbModel.Client().Indexes(name)
+
 	r := &RowCache{
-		name:     name,
-		dbModel:  dbModel,
-		indexes:  newColumnToValue(dbModel.Schema.Table(name).Indexes),
-		dataType: dataType,
-		cache:    make(map[string]model.Model),
-		mutex:    sync.RWMutex{},
+		name:       name,
+		dbModel:    dbModel,
+		indexSpecs: make([]indexSpec, 0, len(schemaIndexes)+len(clientIndexes)),
+		dataType:   dataType,
+		cache:      make(map[string]model.Model),
+		mutex:      sync.RWMutex{},
 	}
+
+	// respect the order of indexes, add first schema indexes, then client
+	// indexes
+	indexes := map[index]indexSpec{}
+	for _, columns := range schemaIndexes {
+		columnKeys := newColumnKeysFromColumns(columns...)
+		index := newIndexFromColumnKeys(columnKeys...)
+		spec := indexSpec{index: index, columns: columnKeys, indexType: schemaIndexType}
+		r.indexSpecs = append(r.indexSpecs, spec)
+		indexes[index] = spec
+	}
+	for _, clientIndex := range clientIndexes {
+		columnKeys := clientIndex.Columns
+		index := newIndexFromColumnKeys(columnKeys...)
+		// if this is already a DB index, ignore
+		if _, ok := indexes[index]; ok {
+			continue
+		}
+		spec := indexSpec{index: index, columns: columnKeys, indexType: clientIndexType}
+		r.indexSpecs = append(r.indexSpecs, spec)
+		indexes[index] = spec
+	}
+
+	r.indexes = r.newIndexes()
 	return r
 }
 
-func newColumnToValue(schemaIndexes [][]string) columnToValue {
-	// RFC 7047 says that Indexes is a [<column-set>] and "Each <column-set> is a set of
-	// columns whose values, taken together within any given row, must be
-	// unique within the table". We'll store the column names, separated by comma
-	// as we'll assume (RFC is not clear), that comma isn't valid in a <id>
-	var indexes []index
-	for i := range schemaIndexes {
-		indexes = append(indexes, newIndex(schemaIndexes[i]...))
-	}
+func (r *RowCache) newIndexes() columnToValue {
 	c := make(columnToValue)
-	for _, index := range indexes {
-		c[index] = make(valueToUUID)
+	for _, indexSpec := range r.indexSpecs {
+		index := indexSpec.index
+		c[index] = make(valueToUUIDs)
 	}
 	return c
 }
@@ -1034,13 +1176,12 @@ func (t *TableCache) ApplyModifications(tableName string, base model.Model, upda
 	return nil
 }
 
-func valueFromIndex(info *mapper.Info, index index) (interface{}, error) {
-	columns := index.columns()
-	if len(columns) > 1 {
+func valueFromIndex(info *mapper.Info, columnKeys []model.ColumnKey) (interface{}, error) {
+	if len(columnKeys) > 1 {
 		var buf bytes.Buffer
 		enc := gob.NewEncoder(&buf)
-		for _, column := range columns {
-			val, err := info.FieldByColumn(column)
+		for _, columnKey := range columnKeys {
+			val, err := valueFromColumnKey(info, columnKey)
 			if err != nil {
 				return "", err
 			}
@@ -1053,9 +1194,37 @@ func valueFromIndex(info *mapper.Info, index index) (interface{}, error) {
 		val := hex.EncodeToString(h.Sum(buf.Bytes()))
 		return val, nil
 	}
-	val, err := info.FieldByColumn(columns[0])
+	val, err := valueFromColumnKey(info, columnKeys[0])
+	if err != nil {
+		return "", err
+	}
+	return val, err
+}
+
+func valueFromColumnKey(info *mapper.Info, columnKey model.ColumnKey) (interface{}, error) {
+	val, err := info.FieldByColumn(columnKey.Column)
 	if err != nil {
 		return nil, err
 	}
+	if columnKey.Key != nil {
+		val, err = valueFromMap(val, columnKey.Key)
+		if err != nil {
+			return "", fmt.Errorf("can't get key value from map: %v", err)
+		}
+	}
 	return val, err
+}
+
+func valueFromMap(aMap interface{}, key interface{}) (interface{}, error) {
+	m := reflect.ValueOf(aMap)
+	if m.Kind() != reflect.Map {
+		return nil, fmt.Errorf("expected map but got %s", m.Kind())
+	}
+	v := m.MapIndex(reflect.ValueOf(key))
+	if !v.IsValid() {
+		// return the zero value for the map value type
+		return reflect.Indirect(reflect.New(m.Type().Elem())).Interface(), nil
+	}
+
+	return v.Interface(), nil
 }