A user may specify additional indexes at runtime through the client DB
model at client initialization. Indexes are specified in sets per table.
Each set consists of columns and optionally a key within the column that
can be addressed if the column type is a map.

The main use case is performance. Running a predicate on all the rows of
a table can be costly. Introducing new indexes in the schema is not backwards
compatible or might not align with the general concept of the schema but those
additional indexes might be useful for specific uses of that schema that 
would otherwise rely on running less performant predicates frequently.

For example, with OVN-K, checking whether ~40 LBs are present in the cache
by a predicate that matches by name on a table that contains ~200K rows took 
around 3 seconds. While surely there is different ways to optimize a predicate, 
for OVN-K which gives & frequently addresses LBs by unique names would 
make sense to have them indexed, something that is not possible to do with 
the schema without breaking backward compatibility.

This can be enhanced in the future with more use cases. For example, supporting
primary and secondary client indexes where primary client indexes could be 
checked for duplicates against a transaction before sending it to the server.

This is implemented in the cache tracking client indexes as if they were
schema indexes. Indexes map to a set of uuids instead of a single uuid
where the set will consist of a single uuid for schema index but might
consist of multiple uuids for client indexes. A new `RowsByModel` method
queries the cache considering schema and client indexes. `RowsByCondition`
also makes use of the new client indexes naturally. The legacy
`RowByModel` method considers only schema indexes. Schema indexes have
priority over client indexes in general.

API wise, client indexes are leveraged through the conditional
APIs from `Where(...)` and `WhereAll(...)`.

So where you would have before something like
```
// slow predicate run on all the LB table rows...
WhereCache(func (lb *LoadBalancer) bool {
    return lb.ExternalIds["myIdKey"] == "myIdValue"
}).List(ctx, &results)
```

Now you can have
```
dbModel, err := nbdb.FullDatabaseModel()
db.SetIndexes(map[string][]model.ClientIndex{
  "Load_Balancer": {{Columns: []model.ColumnKey{Column: "external_ids", Key: "myIdKey"}}}
})

// connect and everything....

lb := &LoadBalancer{
    ExternalIds: map[string]string{"myIdKey": "myIdValue"},
}
// quick indexed result
Where(lb).List(ctx, &results)
```

Other changes in this context:

* Cherry-picked @dave-tucker's [api: Allow use of Where for client-side ops](https://github.com/ovn-org/libovsdb/commit/b54c97c4b95409f779814be465a21f252082554a) 
  so that client indexes can be leveraged through the `Where.List` and `WhereAll.List` 
  conditional APIs. 
* As improvements to the above, run the predicates on shallow copies of model, 
  remove the unnecessary clone and return cache matches directly instead of 
  in-directing through RowsByCondition.
* Fixed an issue with RowsByCondition that was implementing an `Or` between the 
  conditions instead of an `And`. also improved performance.
* Instead of translating model conditions to ovsdb conditions and
  sending those to the server, perform a query on the cache through
  `RowsByModel` or `RowsByCondition` and generate uuid based conditions
  off those results. Fallback to the original conditionals if no cache hits are obtained.

  This is similar to what was already being done for the predicate conditional, this PR
  does the same for the equality and explicit conditionals.

  This has two benefits:

  * OVSDB server does not index schema indexes, it only ensures uniqueness
    but operations based on those indexes are not necessarily fast. By
    querying the cache for the uuids we improve the performance of index
    based operations. Users will no longer be required to do a `Get` before any 
    server side op to ensure this is optimized through the uuid.
  * The conditionals APIs can fully benefit of client indexes for both cache and 
    server side ops.
  
**RFCs**

* The `ClientDBModel`API could be better but I did not want to complicate it 
  unnecessarily. Feedback welcome.
* With this change we rely more on the cache. But the cache doesn't seem to
  be aware at all of monitored vs non-monitored columns and will happily index
  zero values from non-monitored columns. I am not worried in excess about it
  in the context of this PR as this is an issue already present, but perhaps we
  ought to do something about it.