@jcaamano 2022-05-12
Added a couple of commits to benchmark & optimize `RowsByCondition` so that it makes use of indexes as much as possible.

Results before this PR:
```
goos: linux
goarch: amd64
pkg: github.com/ovn-org/libovsdb/cache
cpu: Intel Core Processor (Skylake, IBRS)
BenchmarkRowsByCondition/by_uuid-4         	      38	  28737606 ns/op	 6850409 B/op	  140222 allocs/op
BenchmarkRowsByCondition/by_single_column_squema_index-4         	    1686	    742480 ns/op	     925 B/op	      16 allocs/op
BenchmarkRowsByCondition/by_multi_column_squema_index-4          	      19	  60704196 ns/op	13540612 B/op	  280450 allocs/op
BenchmarkRowsByCondition/by_two_squema_indexes-4                 	     817	   1537409 ns/op	    1516 B/op	      29 allocs/op
BenchmarkRowsByCondition/by_squema_index_and_non-index-4         	      37	  31416413 ns/op	 6772972 B/op	  140246 allocs/op
BenchmarkRowsByCondition/by_single_non_index-4                   	      37	  30643479 ns/op	 6770317 B/op	  140226 allocs/op
BenchmarkRowsByCondition/by_multiple_non_indexes-4               	      19	  60734832 ns/op	13542517 B/op	  280456 allocs/op
PASS
ok  	github.com/ovn-org/libovsdb/cache	8.821s
```

Results before optimization commit (includes using single-column indexes, match all logic and avoids unnecessary clones):
```
goos: linux
goarch: amd64
pkg: github.com/ovn-org/libovsdb/cache
cpu: Intel Core Processor (Skylake, IBRS)
BenchmarkRowsByCondition/by_uuid-4         	  272794	      4057 ns/op	    1181 B/op	      20 allocs/op
BenchmarkRowsByCondition/by_single_column_squema_index-4         	  285651	      4229 ns/op	    1181 B/op	      20 allocs/op
BenchmarkRowsByCondition/by_single_column_client_index-4         	  288646	      4283 ns/op	    1181 B/op	      20 allocs/op
BenchmarkRowsByCondition/by_multi_column_squema_index-4          	     400	   2986806 ns/op	  241852 B/op	   20036 allocs/op
BenchmarkRowsByCondition/by_two_squema_indexes-4                 	  224726	      5065 ns/op	    1645 B/op	      26 allocs/op
BenchmarkRowsByCondition/by_squema_index_and_non-index-4         	  201648	      5581 ns/op	    1749 B/op	      32 allocs/op
BenchmarkRowsByCondition/by_single_non_index-4                   	     387	   3047246 ns/op	  241289 B/op	   20024 allocs/op
BenchmarkRowsByCondition/by_multiple_non_indexes-4               	     392	   2976273 ns/op	  241866 B/op	   20036 allocs/op
PASS
ok  	github.com/ovn-org/libovsdb/cache	11.000s
```

Results after optimization (includes using multi-column indexes):
```
goos: linux
goarch: amd64
pkg: github.com/ovn-org/libovsdb/cache
cpu: Intel Core Processor (Skylake, IBRS)
BenchmarkRowsByCondition/by_uuid-4         	  269812	      4106 ns/op	    1221 B/op	      22 allocs/op
BenchmarkRowsByCondition/by_single_column_squema_index-4         	  198980	      5833 ns/op	    1797 B/op	      36 allocs/op
BenchmarkRowsByCondition/by_single_column_client_index-4         	  178773	      6490 ns/op	    2013 B/op	      45 allocs/op
BenchmarkRowsByCondition/by_multi_column_squema_index-4          	   92376	     12630 ns/op	    5090 B/op	     112 allocs/op
BenchmarkRowsByCondition/by_two_squema_indexes-4                 	  109209	     10330 ns/op	    3957 B/op	      88 allocs/op
BenchmarkRowsByCondition/by_squema_index_and_non-index-4         	  112953	     10330 ns/op	    3957 B/op	      93 allocs/op
BenchmarkRowsByCondition/by_single_non_index-4                   	     397	   3073535 ns/op	  241862 B/op	   20042 allocs/op
BenchmarkRowsByCondition/by_multiple_non_indexes-4               	     391	   2958540 ns/op	  244022 B/op	   20099 allocs/op
PASS
ok  	github.com/ovn-org/libovsdb/cache	10.820s
```

Being able to use multi-column indexes does have a cost for the single-column index performance but I think it makes sense to have it. It's an overall good improvement for a start and I am sure this can be optimized further in various ways. Ideas welcome though if you see easy picks.