@jcaamano 2022-05-12
Pushed an update that shortcuts RowsByCondition when we have reduced the possible index matches to 1. That will help in most situations where I would think we would expect a single match. As said, this can be tuned and optimized further but don't want to go overboard with no experience.

Other than that just added some additional tests and improvement to godoc  that come to mind as times goes on.

Updated results, added a benchmark that uses many conditions.

Results before this PR:
```
goos: linux
goarch: amd64
pkg: github.com/ovn-org/libovsdb/cache
cpu: Intel Core Processor (Skylake, IBRS)
BenchmarkRowsByCondition/by_uuid-4         	      39	  31181645 ns/op	 7156346 B/op	  140223 allocs/op
BenchmarkRowsByCondition/by_single_column_squema_index-4         	    1608	    719762 ns/op	     956 B/op	      16 allocs/op
BenchmarkRowsByCondition/by_multi_column_squema_index-4          	      16	  65336847 ns/op	14151228 B/op	  280448 allocs/op
BenchmarkRowsByCondition/by_two_squema_indexes-4                 	     800	   1582913 ns/op	    1577 B/op	      29 allocs/op
BenchmarkRowsByCondition/by_squema_index_and_non-index-4         	      33	  33603397 ns/op	 7077108 B/op	  140241 allocs/op
BenchmarkRowsByCondition/by_single_non_index-4                   	      32	  32456669 ns/op	 7076734 B/op	  140228 allocs/op
BenchmarkRowsByCondition/by_multiple_non_indexes-4               	      18	  64228400 ns/op	14151768 B/op	  280449 allocs/op
BenchmarkRowsByCondition/by_many_conditions-4                    	       8	 141383613 ns/op	33114615 B/op	  600961 allocs/op
PASS
ok  	github.com/ovn-org/libovsdb/cache	9.916s
```

Results before optimization commit (includes using single-column indexes, match all logic and avoids unnecessary clones):
```
goos: linux
goarch: amd64
pkg: github.com/ovn-org/libovsdb/cache
cpu: Intel Core Processor (Skylake, IBRS)
BenchmarkRowsByCondition/by_uuid-4         	  207000	      5366 ns/op	    1854 B/op	      32 allocs/op
BenchmarkRowsByCondition/by_single_column_squema_index-4         	  215368	      5775 ns/op	    1854 B/op	      32 allocs/op
BenchmarkRowsByCondition/by_single_column_client_index-4         	     166	   6692064 ns/op	 1842801 B/op	   50041 allocs/op
BenchmarkRowsByCondition/by_multi_column_squema_index-4          	     370	   2896103 ns/op	  242527 B/op	   20048 allocs/op
BenchmarkRowsByCondition/by_two_squema_indexes-4                 	  161007	      6484 ns/op	    2318 B/op	      38 allocs/op
BenchmarkRowsByCondition/by_squema_index_and_non-index-4         	  154776	      6872 ns/op	    2428 B/op	      44 allocs/op
BenchmarkRowsByCondition/by_single_non_index-4                   	     387	   3063313 ns/op	  241962 B/op	   20036 allocs/op
BenchmarkRowsByCondition/by_multiple_non_indexes-4               	     396	   2947208 ns/op	  242533 B/op	   20048 allocs/op
BenchmarkRowsByCondition/by_many_conditions-4                    	   89371	     12340 ns/op	    5422 B/op	      93 allocs/op
PASS
ok  	github.com/ovn-org/libovsdb/cache	13.949s
```

Results after optimization (includes using multi-column indexes):
```
goos: linux
goarch: amd64
pkg: github.com/ovn-org/libovsdb/cache
cpu: Intel Core Processor (Skylake, IBRS)
BenchmarkRowsByCondition/by_uuid-4         	  208352	      5577 ns/op	    1894 B/op	      34 allocs/op
BenchmarkRowsByCondition/by_single_column_squema_index-4         	  160498	      7098 ns/op	    2428 B/op	      47 allocs/op
BenchmarkRowsByCondition/by_single_column_client_index-4         	  111760	      9768 ns/op	    3677 B/op	      75 allocs/op
BenchmarkRowsByCondition/by_multi_column_squema_index-4          	   79140	     14264 ns/op	    5667 B/op	     123 allocs/op
BenchmarkRowsByCondition/by_two_squema_indexes-4                 	  149360	      7876 ns/op	    2910 B/op	      54 allocs/op
BenchmarkRowsByCondition/by_squema_index_and_non-index-4         	  128370	      7967 ns/op	    2910 B/op	      54 allocs/op
BenchmarkRowsByCondition/by_single_non_index-4                   	     390	   3027351 ns/op	  242536 B/op	   20054 allocs/op
BenchmarkRowsByCondition/by_multiple_non_indexes-4               	     403	   2975605 ns/op	  244693 B/op	   20111 allocs/op
BenchmarkRowsByCondition/by_many_conditions-4                    	   95650	     12171 ns/op	    5694 B/op	      89 allocs/op
PASS
ok  	github.com/ovn-org/libovsdb/cache	12.210s
```

Was curious to have a benchmark of index vs predicate. It turns out there was one for predicates but I don't it was working fine (too good to be true) so I updated it. Results: 
```
BenchmarkAPIListPredicate/ApiListPredicate:_predicate_returns_none-4         	     363	   3117351 ns/op	  930387 B/op	   20010 allocs/op
BenchmarkAPIListPredicate/ApiListPredicate:_predicate_returns_all-4          	     225	   5167648 ns/op	 2445557 B/op	   30223 allocs/op
BenchmarkAPIListPredicate/ApiListPredicate:_predicate_on_an_arbitrary_condition-4         	     332	   3627668 ns/op	 1123775 B/op	   21152 allocs/op
BenchmarkAPIListPredicate/ApiListPredicate:_predicate_matches_name-4                      	     363	   3292371 ns/op	  930691 B/op	   20012 allocs/op
BenchmarkAPIListPredicate/ApiListPredicate:_by_index,_no_predicate-4                      	   45225	     25121 ns/op	   84076 B/op	      32 allocs/op
```